Permisos de Usuario, condiciones de acceso, tabla de permisos

+++ blog/views.py
@@ -4,8 +4,18 @@ from .models import Posts
 from django.contrib.auth.models import User
 from django.conf import settings
 from django.views.decorators.cache import never_cache
-# Create your views here.
+from django.contrib import messages
+from django.contrib.auth.decorators import permission_required
+
+
 def createPage(request):
+
+
+    if(request.user.has_perm('blog.can_view')):
+
+        messages.error(request, 'No tienes permisos para realizar esta operación')
+        return redirect('pages')
+
 
     if(request.method == "POST"):
         
@@ -38,33 +48,47 @@ def pagesListView(request):
 def pageDetailView(request,id):
 
     print(id)
+    print("ver: ",request.user.has_perm('blog.can_view'))
+    print(request.user.username)
+
+
     post = Posts.objects.filter(id=id)
     return render(request,'blog/pageDetails.html',{
         'post' : post
     })
 
-@never_cache
+
 def pageEdit(request,id):
 
+    user=request.user
     posts = Posts.objects.get(id=id)
-    print(posts.title)
-  
+   
+    
+    if (user.has_perm('blog.can_view')):
+        messages.error(request, 'No tienes permisos para realizar esta operación')
+        return redirect('details', id)
+   
     if (request.method == 'GET'):
             return render(request, "blog/pageEdit.html",{
             'form': PostsEditForm(instance=posts)
             })
    
     else:
-         
+        
         form = PostsEditForm(request.POST,request.FILES, instance=posts)
         print(form)
         if form.is_valid():
-            post_instance = form.save(commit=False)
-        
-            if form.cleaned_data.get('delete_image'):
-                if (post_instance.imageMain):
-                    post_instance.imageMain.delete()
-                    post_instance.imageMain = None
+            
+            
+            if(user.has_perm('blog.can_edit')):
+                messages.error(request, 'Access Denied - No tienes permisos para cambiar la imagen')
+                return render(request,'blog/pageEdit.html',{'form': PostsEditForm(instance=posts)})
+            else:
+                post_instance = form.save(commit=False)
+                if form.cleaned_data.get('delete_image'):
+                    if (post_instance.imageMain):
+                        post_instance.imageMain.delete()
+                        post_instance.imageMain = None
             
             # Guardar los cambios en la publicación
             post_instance.save()
@@ -76,10 +100,17 @@ def pageEdit(request,id):
         'form': PostsEditForm(instance=posts)
     })
 
+
 def deletePage(request,id):
 
-    
-    post = Posts.objects.get(id=id)
-    post.delete()
+    user = request.user
+    if(user.has_perm('blog.can_view') or (user.has_perm('blog.can_edit'))):
+  
+        messages.error(request, 'No tienes permisos para realizar esta operación')
+        return redirect('details', id)
+    else:
+        post = Posts.objects.get(id=id)
+        post.delete()
 
     return redirect('pages')
+
